+++ src/app/lib/lforms-widget-registry.ts
@@ -47,6 +47,7 @@ import { ExpressionEditorComponent } from './widgets/expression-editor/expressio
 import { ExpressionEditorDlgComponent } from './widgets/expression-editor-dlg/expression-editor-dlg.component';
 import { VariableComponent } from './widgets/variable/variable.component';
 import { InitialNumberComponent } from './widgets/initial-number/initial-number.component';
+import { AnswerValueSetCodingDisplayComponent } from './widgets/answer-value-set-coding-display/answer-value-set-coding-display.component';
 
 @Injectable()
 export class LformsWidgetRegistry extends DefaultWidgetRegistry {
@@ -94,5 +95,6 @@ export class LformsWidgetRegistry extends DefaultWidgetRegistry {
     this.register('expression-editor-dlg', ExpressionEditorDlgComponent);
     this.register('variable', VariableComponent);
     this.register('initial-number', InitialNumberComponent);
+    this.register('answer-value-set-coding-display', AnswerValueSetCodingDisplayComponent);
   }
 }