+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project follows [Semantic Versioning](http://semver.org/).
 ## [9.7.2] 2024-12-21
 ### Added
 - Added Value Method field for selecting initial values.
+- Added initial values for SNOMED and non-SNOMED answer value sets.
 
 ## [9.7.1] 2024-12-20
 ### Changed